Update and rename cmd_requirements.py to position.py
